made changes to test and course.JS
