+++ test/classes/CourseLookupTest.js
@@ -1,0 +1,32 @@
+var expect = require('chai').expect;
+var assert = require('chai').assert;
+var CourseJS = require('../../course.js');
+
+// CourseLookup test
+describe('CourseLookup', function() {
+
+    // constructor test
+    describe('#CourseLookup()', function() {
+        //TODO:
+    });
+
+    // insert test
+    describe('#insert()', function() {
+        //TODO: Implement Test
+    });
+
+    // getCourseByAlias test
+    describe('#getCourseByAlias()', function() {
+        //TODO: Implement Test
+    });
+
+    // findOtherSectionsOfCourse test
+    describe('#findOtherSectionsOfCourse()', function() {
+        //TODO: Implement Test
+    });
+
+    // findMatchingCourses test
+    describe('#findMatchingCourses()', function() {
+        //TODO: Implement Test
+    });
+});
