+++ test/classes/EntryGroupTest.js
@@ -1,0 +1,57 @@
+var expect = require('chai').expect;
+var assert = require('chai').assert;
+var CourseJS = require('../../course.js');
+
+// EntryGroup Tests
+describe('EntryGroup', function() {
+
+    // constructor Test
+    describe('#EntryGroup()', function() {
+        //TODO: Implement Test
+    });
+
+    // insert test
+    describe('#insert()', function() {
+        //TODO: Implement Test
+    });
+
+    // select test
+    describe('#select()', function() {
+        //TODO: Implement Test
+    });
+
+    // activate test
+    describe('#activate()', function() {
+        //TODO: Implement Test
+    });
+
+    // deactivate test
+    describe('#deactivate()', function() {
+        //TODO: Implement Test
+    });
+
+    // getSelectedEntry test
+    describe('#getSelectedEntry()', function() {
+        //TODO: Implement Test
+    });
+
+    // getActivatedEntries test
+    describe('#getActivatedEntries()', function() {
+        //TODO: Implement Test
+    });
+
+    // getOverlappingTimeSets test
+    describe('#getOverlappingTimeSets()', function() {
+        //TODO: Implement Test
+    });
+
+    // isCompatibleWithEntryGroup test
+    describe('#isCompatibleWithEntryGroup()', function() {
+        //TODO: Implement Test
+    });
+
+    // getInfo test
+    describe('#getInfo()', function() {
+        //TODO: Implement Test
+    });
+});
