+++ test/classes/ScheduleTest.js
@@ -1,0 +1,32 @@
+var expect = require('chai').expect;
+var assert = require('chai').assert;
+var CourseJS = require('../../course.js');
+
+// Schedule Tests
+describe('Schedule', function() {
+
+    // constructor test
+    describe('#Schedule()', function() {
+        //TODO: Implement Test
+    });
+
+    // isCompatibleWithSchedule test
+    describe('#isCompatibleWithSchedule()', function() {
+        //TODO: Implement Test
+    });
+
+    // getItems test
+    describe('#getItems()', function() {
+        //TODO: Implement Test
+    });
+
+    // getItemsForDay test
+    describe('#getItemsForDay()', function() {
+        //TODO: Implement Test
+    });
+
+    // getFreeTime test
+    describe('#getFreeTime()', function() {
+        //TODO: Implement Test
+    });
+});
