+++ test/classes/SearchQueryTest.js
@@ -1,0 +1,17 @@
+var expect = require('chai').expect;
+var assert = require('chai').assert;
+var CourseJS = require('../../course.js');
+
+// SearchQuery test
+describe('SearchQuery', function() {
+
+    // constructor test
+    describe('#searchQuery()', function() {
+        //TODO: Implement Test
+    });
+
+    // formatData test
+    describe('#formatData()', function() {
+        //TODO: Implement Test
+    });
+});
